+++ src/kernel.rs
@@ -1,256 +1,13 @@
-use std::{borrow::Cow, path::Path};
+use crate::*;
 
-use crate::{
-    primitives::{BufOps, ImgOps, PixelInfo},
-    DescriptorSet, Framework, GpuBuffer, GpuBufferUsage, GpuConstImage, GpuImage, GpuUniformBuffer,
-    Kernel, Program, Shader,
-};
-
-impl<'res> DescriptorSet<'res> {
-    /// Binds a [`GpuUniformBuffer`] as a uniform buffer in the shader.
-    ///
-    /// ### Example WGSL syntax:
-    /// ```ignore
-    /// struct UniformStruct {
-    ///     a: vec3<u32>;
-    ///     b: vec3<u32>;
-    ///     c: vec3<u32>;
-    /// };
-    ///
-    /// [[group(0), binding(0)]]
-    /// var<uniform> myUniformBuffer: UniformStruct;
-    /// ```
-    ///
-    /// ### Example GLSL syntax:
-    /// ```glsl
-    /// layout(std140, binding = 0)
-    /// uniform UniformStruct {
-    ///     uvec3 a;
-    ///     uvec3 b;
-    ///     uvec3 c;
-    /// };
-    /// ```
-    pub fn bind_uniform_buffer<T>(
-        mut self,
-        uniform_buf: &'res GpuUniformBuffer<T>,
-        bind_id: u32,
-    ) -> Self
-    where
-        T: bytemuck::Pod,
-    {
-        let bind_entry = wgpu::BindGroupLayoutEntry {
-            binding: bind_id,
-            visibility: wgpu::ShaderStages::COMPUTE,
-            ty: wgpu::BindingType::Buffer {
-                has_dynamic_offset: false,
-                min_binding_size: None,
-                ty: wgpu::BufferBindingType::Uniform,
-            },
-            count: None,
-        };
-
-        let bind = wgpu::BindGroupEntry {
-            binding: bind_id,
-            resource: uniform_buf.as_binding_resource(),
-        };
-
-        self.set_layout.push(bind_entry);
-        self.binds.push(bind);
-
-        self
-    }
-
-    /// Binds a [`GpuBuffer`] as a storage buffer in the shader with a specific `usage`.
-    ///
-    /// ### Example WGSL syntax:
-    /// ```ignore
-    /// struct StorageStruct {
-    ///     data: [[stride(4)]] array<i32>;
-    /// };
-    ///
-    /// [[group(0), binding(0)]]
-    /// var<storage, read_write> myStorageBuffer: StorageStruct;
-    /// ```
-    ///
-    /// ### Example GLSL syntax:
-    /// ```glsl
-    /// layout (set=0, binding=0) buffer myStorageBuffer {
-    ///     int data[];
-    /// };
-    /// ```
-    pub fn bind_buffer<T>(
-        mut self,
-        storage_buf: &'res GpuBuffer<T>,
-        usage: GpuBufferUsage,
-        bind_id: u32,
-    ) -> Self
-    where
-        T: bytemuck::Pod,
-    {
-        let bind_entry = wgpu::BindGroupLayoutEntry {
-            binding: bind_id,
-            visibility: wgpu::ShaderStages::COMPUTE,
-            ty: wgpu::BindingType::Buffer {
-                has_dynamic_offset: false,
-                min_binding_size: None,
-                ty: wgpu::BufferBindingType::Storage {
-                    read_only: usage == GpuBufferUsage::ReadOnly,
-                },
-            },
-            count: None,
-        };
-
-        let bind = wgpu::BindGroupEntry {
-            binding: bind_id,
-            resource: storage_buf.as_binding_resource(),
-        };
-
-        self.set_layout.push(bind_entry);
-        self.binds.push(bind);
-
-        self
-    }
-
-    /// Binds a [`GpuImage`] as a storage image in the shader.
-    /// This image is write-only.
-    /// ### Example WGSL syntax:
-    /// ```ignore
-    /// [[group(0), binding(0)]]
-    /// var myStorageImg: texture_storage_2d<rgba8uint, write>;
-    /// ```
-    ///
-    /// ### Example GLSL syntax:
-    /// ```glsl
-    /// layout (set=0, binding=0, rgba8uint) uimage2D myStorageImg;
-    /// ```
-    pub fn bind_image<P: PixelInfo>(mut self, img: &'res GpuImage<P>, bind_id: u32) -> Self {
-        let bind_entry = wgpu::BindGroupLayoutEntry {
-            binding: bind_id,
-            visibility: wgpu::ShaderStages::COMPUTE,
-            ty: wgpu::BindingType::StorageTexture {
-                access: wgpu::StorageTextureAccess::WriteOnly,
-                format: P::wgpu_format(),
-                view_dimension: wgpu::TextureViewDimension::D2,
-            },
-            count: None,
-        };
-
-        let bind = wgpu::BindGroupEntry {
-            binding: bind_id,
-            resource: img.as_binding_resource(),
-        };
-
-        self.set_layout.push(bind_entry);
-        self.binds.push(bind);
-
-        self
-    }
-
-    /// Binds a [`GpuConstImage`] as a texture in the shader.
-    /// This image is read-only.
-    /// ### Example WGSL syntax:
-    /// ```ignore
-    /// [[group(0), binding(0)]]
-    /// var myTexture: texture_2d<u32>;
-    /// ```
-    ///
-    /// ### Example GLSL syntax:
-    /// ```glsl
-    /// layout (set=0, binding=0) utexture2D myTexture;
-    /// ```
-    pub fn bind_const_image<P>(mut self, img: &'res GpuConstImage<P>, bind_id: u32) -> Self
-    where
-        P: PixelInfo,
-    {
-        let bind_entry = wgpu::BindGroupLayoutEntry {
-            binding: bind_id,
-            visibility: wgpu::ShaderStages::COMPUTE,
-            ty: wgpu::BindingType::Texture {
-                sample_type: P::wgpu_texture_sample(),
-                multisampled: false,
-                view_dimension: wgpu::TextureViewDimension::D2,
-            },
-            count: None,
-        };
-
-        let bind = wgpu::BindGroupEntry {
-            binding: bind_id,
-            resource: img.as_binding_resource(),
-        };
-
-        self.set_layout.push(bind_entry);
-        self.binds.push(bind);
-
-        self
-    }
-}
-
-impl Shader {
-    /// Initialises a [`Shader`] from a SPIR-V file.
-    pub fn from_spirv_file(fw: &Framework, path: impl AsRef<Path>) -> std::io::Result<Self> {
-        let bytes = std::fs::read(&path)?;
-        let shader_name = path.as_ref().to_str();
-
-        Ok(Self::from_spirv_bytes(fw, &bytes, shader_name))
-    }
-
-    /// Initialises a [`Shader`] from SPIR-V bytes with an optional `name`.
-    pub fn from_spirv_bytes(fw: &Framework, bytes: &[u8], name: Option<&str>) -> Self {
-        let source = wgpu::util::make_spirv(bytes);
-
-        let shader = fw
-            .device
-            .create_shader_module(wgpu::ShaderModuleDescriptor {
-                label: name,
-                source,
-            });
-
-        Self(shader)
-    }
-
-    /// Initialises a [`Shader`] from a `WGSL` file.
-    pub fn from_wgsl_file(fw: &Framework, path: impl AsRef<Path>) -> std::io::Result<Self> {
-        let source_string = std::fs::read_to_string(&path)?;
-        let shader_name = path.as_ref().to_str();
-
-        Ok(Self(fw.device.create_shader_module(
-            wgpu::ShaderModuleDescriptor {
-                label: shader_name,
-                source: wgpu::ShaderSource::Wgsl(Cow::Owned(source_string)),
-            },
-        )))
-    }
-
-    /// Initialises a [`Shader`] from a `WGSL` string.
-    pub fn from_wgsl_string(
-        fw: &Framework,
-        source: String,
-        name: Option<&str>,
-    ) -> std::io::Result<Self> {
-        Ok(Self(fw.device.create_shader_module(
-            wgpu::ShaderModuleDescriptor {
-                label: name,
-                source: wgpu::ShaderSource::Wgsl(Cow::Owned(source)),
-            },
-        )))
-    }
-}
-
-impl<'sha, 'res> Program<'sha, 'res> {
-    /// Creates a new [`Program`] using a `shader` and an `entry_point`.
-    pub fn new(shader: &'sha Shader, entry_point: impl Into<String>) -> Self {
-        Self {
-            shader,
-            entry_point: entry_point.into(),
-            descriptors: Vec::new(),
-        }
-    }
-
-    /// Adds a [`DescriptorSet`] to this [`Program`] layout.
-    pub fn add_descriptor_set(mut self, desc: DescriptorSet<'res>) -> Self {
-        self.descriptors.push(desc);
-        self
-    }
+/// Used to enqueue the execution of a shader with the bidings provided.
+///
+/// Equivalent to OpenCL's Kernel.
+pub struct Kernel<'fw> {
+    fw: &'fw Framework,
+    pipeline: wgpu::ComputePipeline,
+    sets: Vec<wgpu::BindGroup>,
+    entry_point: String,
 }
 
 impl<'fw> Kernel<'fw> {
